Further case fixes.

+++ trunk/sdmvis/e7/Engines/Trackball2.cpp
@@ -92,7 +92,7 @@ void Trackball2::update( float ax, float ay, float bx, float by, int mode )
 		glm::vec3 R = glm::normalize( glm::cross(a,b) );
 		
 		// Unit quaternion for this rotation
-		glm::fquat q( cos(theta/2.f), R*sin(theta/2.f) );
+        glm::fquat q( cos(theta/2.f), R*(float)sin(theta/2.f) );
 		
 		if( m_immediateUpdate )
 		{
